Move deprecation to own file and create deprecation for removing maxWidth from save
